Move InputFlow to separate class.
